requirements.txt
hydra-core>=1.1.1
pandas>=1.3.4
numpy>=1.19.5
torch>=1.10.0
torchvision>=0.11.1
tensorboard>=2.7.0
pytorch-lightning>=1.5.3
torchsummary>=.5.1
Pillow>=8.4.0